almost done, test with other testers

+++ count/count_digit.c
@@ -6,16 +6,15 @@
 /*   By: yoav <yoav@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/20 13:59:32 by yoav              #+#    #+#             */
-/*   Updated: 2022/06/23 16:39:51 by yoav             ###   ########.fr       */
+/*   Updated: 2022/06/26 18:42:19 by yoav             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-size_t	count_digit_by_base(long n, size_t base)
+size_t	count_digit_by_base(long n, long base)
 {
 	size_t	ret;
-
 	ret = 0;
 	if (n == 0)
 		return (1);
